[ALPS04729707] PTPOD: Modify data type code defect
[Detail] PTPOD: Modify data type code defect

MTK-Commit-Id: 142e0a536751160d20d0bd15fd63fcc3ce119ce9

CR-Id: ALPS04729707
Feature: PTP-OD (Performance Thermal Power-Over Drive)

+++ drivers/misc/mediatek/base/power/upower_v2/mtk_unified_power_mt6763.c
@@ -222,7 +222,8 @@ static void upower_scale_l_cap(void)
 			tbl = upower_tbl_infos[i].p_upower_tbl;
 			for (j = 0; j < UPOWER_OPP_NUM; j++) {
 				temp = tbl->row[j].cap;
-				tbl->row[j].cap = temp * ratio / 1000;
+				tbl->row[j].cap = (unsigned long long)temp *
+					ratio / 1000;
 			}
 		}
 	}
